+++ wrapper/Convert/SireOpenMM/sire_to_openmm_system.cpp
@@ -230,8 +230,6 @@ void _add_bond_restraints(const SireMM::BondRestraints &restraints,
     const double internal_to_nm = (1 * SireUnits::angstrom).to(SireUnits::nanometer);
     const double internal_to_k = (1 * SireUnits::kcal_per_mol / (SireUnits::angstrom2)).to(SireUnits::kJ_per_mol / (SireUnits::nanometer2));
 
-    auto cljff = lambda_lever.getForce<OpenMM::NonbondedForce>("clj", system);
-
     std::vector<double> custom_params = {1.0, 0.0, 0.0};
 
     for (const auto &restraint : atom_restraints)
@@ -307,8 +305,6 @@ void _add_morse_potential_restraints(const SireMM::MorsePotentialRestraints &res
     const double internal_to_nm = (1 * SireUnits::angstrom).to(SireUnits::nanometer);
     const double internal_to_k = (1 * SireUnits::kcal_per_mol / (SireUnits::angstrom2)).to(SireUnits::kJ_per_mol / (SireUnits::nanometer2));
     const double internal_to_de = (1 * SireUnits::kcal_per_mol).to(SireUnits::kJ_per_mol);
-
-    auto cljff = lambda_lever.getForce<OpenMM::NonbondedForce>("clj", system);
 
     std::vector<double> custom_params = {1.0, 0.0, 0.0, 0.0};
 
@@ -797,7 +793,9 @@ void _set_clj_cutoff(OpenMM::NonbondedForce &cljff,
  */
 std::shared_ptr<std::vector<OpenMM::Vec3>>
 _set_box_vectors(OpenMM::System &system,
-                 const ForceFieldInfo &ffinfo)
+                 const SelectorMol &mols,
+                 const ForceFieldInfo &ffinfo,
+                 const PropertyMap &map)
 {
     // create the periodic box vectors
     std::shared_ptr<std::vector<OpenMM::Vec3>> boxvecs;
@@ -843,6 +841,45 @@ _set_box_vectors(OpenMM::System &system,
             boxvecs_data[1] = OpenMM::Vec3(yx, yy, yz);
             boxvecs_data[2] = OpenMM::Vec3(zx, zy, zz);
         }
+
+        system.setDefaultPeriodicBoxVectors(boxvecs_data[0],
+                                            boxvecs_data[1],
+                                            boxvecs_data[2]);
+    }
+    else
+    {
+        // Set the box vectors based on the AABox of the system and nonbonded
+        // cutoff distance.
+
+        QVector<QVector<SireMaths::Vector>> all_coords;
+
+        // Get the coordinates from all of the molecules.
+        for (const auto &mol : mols)
+        {
+            const auto coords = mol.property(map["coordinates"]).asA<SireMol::AtomCoords>().toVector();
+            all_coords.append(coords);
+        }
+
+        // Create an AABox.
+        const auto aabox = SireVol::AABox(all_coords);
+
+        // Work out the minimum box size and convert from Angstroms to nm.
+        auto min_box = 2.0 * aabox.halfExtents() * 0.1;
+
+        // Adjust the box size based on the cutoff distance.
+        if (ffinfo.hasCutoff())
+        {
+            // Get the nonbonded cutoff.
+            const auto double_cutoff = 2.0 * ffinfo.cutoff().to(SireUnits::nanometers);
+            min_box += SireMaths::Vector(double_cutoff, double_cutoff, double_cutoff);
+        }
+
+        boxvecs.reset(new std::vector<OpenMM::Vec3>(3));
+        auto boxvecs_data = boxvecs->data();
+
+        boxvecs_data[0] = OpenMM::Vec3(min_box.x(), 0, 0);
+        boxvecs_data[1] = OpenMM::Vec3(0, min_box.y(), 0);
+        boxvecs_data[2] = OpenMM::Vec3(0, 0, min_box.z());
 
         system.setDefaultPeriodicBoxVectors(boxvecs_data[0],
                                             boxvecs_data[1],
@@ -948,7 +985,7 @@ OpenMMMetaData SireOpenMM::sire_to_openmm_system(OpenMM::System &system,
     }
 
     // set the box vectors for periodic spaces
-    auto boxvecs = _set_box_vectors(system, ffinfo);
+    auto boxvecs = _set_box_vectors(system, mols, ffinfo, map);
 
     // Are any of the molecules perturbable, and if they are,
     // should we just ignore perturbations?